# openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Home","operationId":"home__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/words":{"get":{"summary":"Predict Words","operationId":"predict_words_words_get","parameters":[{"required":true,"schema":{"title":"Q","type":"string"},"name":"q","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/images":{"get":{"summary":"Search Image","operationId":"search_image_images_get","parameters":[{"required":true,"schema":{"title":"Q","type":"string"},"name":"q","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/distance":{"get":{"summary":"Predict Words","operationId":"predict_words_distance_get","parameters":[{"required":true,"schema":{"title":"Word A","type":"string"},"name":"word_a","in":"query"},{"required":true,"schema":{"title":"Word B"},"name":"word_b","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DistanceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"DistanceResponse":{"title":"DistanceResponse","required":["word_a","word_b","cosine_similarity"],"type":"object","properties":{"word_a":{"title":"Word A","type":"string"},"word_b":{"title":"Word B","type":"string"},"cosine_similarity":{"title":"Cosine Similarity","type":"number"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ImageResponse":{"title":"ImageResponse","required":["word","urls"],"type":"object","properties":{"word":{"title":"Word","type":"string"},"urls":{"title":"Urls","type":"array","items":{"type":"string"}}}},"PredictResponse":{"title":"PredictResponse","required":["words"],"type":"object","properties":{"words":{"title":"Words","type":"array","items":{"$ref":"#/components/schemas/Word"}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}},"Word":{"title":"Word","required":["word","cosine_similarity"],"type":"object","properties":{"word":{"title":"Word","type":"string"},"cosine_similarity":{"title":"Cosine Similarity","type":"number"}}}}}}